extract presenter structure
