+++ app/src/main/java/com/example/florent/stateexperiment/core/arch/PresenterView.kt
@@ -1,0 +1,3 @@
+package com.example.florent.stateexperiment.core.arch
+
+interface PresenterView
