+++ app/src/main/java/com/example/florent/stateexperiment/main/MainActivity.kt
@@ -6,10 +6,12 @@ import android.support.v7.app.AppCompatActivity
 import android.view.View.GONE
 import android.view.View.VISIBLE
 import com.example.florent.stateexperiment.R.layout
-import com.example.florent.stateexperiment.main.MainAction.Refresh
+import com.example.florent.stateexperiment.main.messages.MainAction.Refresh
 import com.example.florent.stateexperiment.main.MainPresenter.View
-import com.example.florent.stateexperiment.main.MainUiModel.Display
-import com.example.florent.stateexperiment.main.MainUiModel.Refreshing
+import com.example.florent.stateexperiment.main.messages.MainUiModel.Display
+import com.example.florent.stateexperiment.main.messages.MainUiModel.Refreshing
+import com.example.florent.stateexperiment.main.messages.MainAction
+import com.example.florent.stateexperiment.main.messages.MainUiModel
 import com.jakewharton.rxbinding2.view.RxView
 import io.reactivex.Observable
 import kotlinx.android.synthetic.main.activity_main.*
@@ -25,7 +27,7 @@ class MainActivity : AppCompatActivity(), View {
         setContentView(layout.activity_main)
 
         ViewModelProviders.of(this)
-                .get(MainViewModel::class.java)
+                .get(MainRotationStateHolder::class.java)
                 .mainComponent
                 .inject(this)
