+++ app/src/main/java/com/example/florent/stateexperiment/main/messages/MainAction.kt
@@ -1,0 +1,5 @@
+package com.example.florent.stateexperiment.main.messages
+
+sealed class MainAction {
+    object Refresh : MainAction()
+}
