+++ app/src/main/java/com/example/florent/stateexperiment/main/messages/MainUiModel.kt
@@ -1,0 +1,6 @@
+package com.example.florent.stateexperiment.main.messages
+
+sealed class MainUiModel {
+    object Refreshing : MainUiModel()
+    data class Display(val name: String) : MainUiModel()
+}
